Potential fix for code scanning alert no. 120

+++ drivers/ddr/marvell/a38x/ddr3_training_ip_engine.c
@@ -1270,7 +1270,7 @@ int ddr3_tip_ip_training_wrapper(u32 dev_num, enum hws_access_type access_type,
 	int status;
 	u32 bit_lock_result;
 
-	u8 sybphy_id;
+	u32 sybphy_id;
 	u32 octets_per_if_num = ddr3_tip_dev_attr_get(dev_num, MV_ATTR_OCTET_PER_INTERFACE);
 	struct mv_ddr_topology_map *tm = mv_ddr_topology_map_get();
